Move `prepare_calldata` to `solidity-verifiers`
