+++ examples/full_flow.rs
@@ -22,10 +22,7 @@ use std::time::Instant;
 use folding_schemes::{
     commitment::{kzg::KZG, pedersen::Pedersen},
     folding::{
-        nova::{
-            decider_eth::{prepare_calldata, Decider as DeciderEth},
-            Nova, PreprocessorParam,
-        },
+        nova::{decider_eth::Decider as DeciderEth, Nova, PreprocessorParam},
         traits::CommittedInstanceOps,
     },
     frontend::FCircuit,
@@ -33,7 +30,7 @@ use folding_schemes::{
     Decider, Error, FoldingScheme,
 };
 use solidity_verifiers::calldata::{
-    get_function_selector_for_nova_cyclefold_verifier, NovaVerificationMode,
+    get_function_selector_for_nova_cyclefold_verifier, prepare_calldata, NovaVerificationMode,
 };
 use solidity_verifiers::{
     evm::{compile_solidity, Evm},
